backend/requirements.txt
flask-graphql
graphene
graphql-relay
graphql-server-core
graphql-core
python-dotenv
Flask 
SQLAlchemy
graphene-sqlalchemy
psycopg2 
flask-sqlalchemy
flask-migrate
